+++ android/src/main/java/com/rnfingerprint/FingerprintHandler.java
@@ -40,7 +40,7 @@ public class FingerprintHandler extends FingerprintManager.AuthenticationCallbac
 
     @Override
     public void onAuthenticationFailed() {
-        mCallback.onError("Not recognized. Try again.", FingerprintAuthConstants.AUTHENTICATION_FAILED);
+        mCallback.onAuthFailed(FingerprintAuthConstants.AUTHENTICATION_FAILED);
     }
 
     @Override
@@ -61,6 +61,8 @@ public class FingerprintHandler extends FingerprintManager.AuthenticationCallbac
         void onAuthenticated();
 
         void onError(String errorString, int errorCode);
+
+        void onAuthFailed(int errorCode);
 
         void onCancelled();
     }